Correción función de buscar ImagenSearcher
